Added Tap to flip Indication and availability of book
Added flip icon to top right of book icon in home page to indicate that icon is flippable. I also added a tag next to the listed by user to indicate if the book is available to borrow or not.
Please Note: Styling and stuff is not final. I just added functionality.